[DI] Refactor integration tests (#4817)
