+++ integration-tests/debugger/target-app/basic.js
@@ -1,0 +1,18 @@
+'use strict'
+
+require('dd-trace/init')
+const Fastify = require('fastify')
+
+const fastify = Fastify()
+
+fastify.get('/:name', function handler (request) {
+  return { hello: request.params.name } // BREAKPOINT
+})
+
+fastify.listen({ port: process.env.APP_PORT }, (err) => {
+  if (err) {
+    fastify.log.error(err)
+    process.exit(1)
+  }
+  process.send({ port: process.env.APP_PORT })
+})
